Moved components closer to related components
