+++ dev/src/components/manual-grading/ReviewInfo.tsx
@@ -1,7 +1,7 @@
 import React from "react"
-import { NoSubmission } from "../consts"
-import { SubmissionStatus } from "../Helpers"
-import { useAppState } from "../overmind"
+import { NoSubmission } from "../../consts"
+import { SubmissionStatus } from "../../Helpers"
+import { useAppState } from "../../overmind"
 
 const ReviewInfo = (): JSX.Element => {
     const { review: { currentReview, reviewer, graded, criteriaTotal }, activeSubmissionLink } = useAppState()
